Selecting campaign
Added a whole system that let you select an campaign in the campaigns
view which will bring you to the campaign detail page.

If you got here from the campaigns view, it will receive the campaign
from the service, if not the website will download it agin in order to
let people save the link for quick acces.